run setup and push as shell script
