+++ SETUP_GUIDE.md
@@ -1,0 +1,20 @@
+# Time syncing Toggl - Tempo
+
+## Prerequisite
+
+- Install node.js: https://nodejs.org/en/download
+- Download sources
+- Setup CLI Tool by running following command from a terminal inside a specific `dist` folder:
+
+```
+node setup.js
+```
+
+- or execute `dist/setup.sh` script and follow the instructions
+
+## Publish Toggl to Tempo
+
+### for today
+
+Run `dist/push-today.sh`
+
