+++ assets/push-today.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+node "$(dirname "$0")/push.js" "$(date +'%Y-%m-%d')"
