+++ assets/setup.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+node "$(dirname "$0")/setup.js"
