+++ package.json
@@ -28,6 +28,7 @@
     "node": ">=18.0.0"
   },
   "devDependencies": {
+    "copy-webpack-plugin": "^12.0.2",
     "webpack": "^5.97.1",
     "webpack-cli": "^6.0.1",
     "webpack-node-externals": "^3.0.0"
